Adding error files for future dialog boxes. Altering the services and adding CORS Bean to allow access to 8080 from 4200

+++ com.omni.systems.mailinglist.service/src/main/java/application/Subscribers.java
@@ -9,15 +9,15 @@ import javax.persistence.Id;
 public class Subscribers {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private Integer subscriber_id;
+	private Long subscriber_id;
 	private String first_name;
 	private String last_name;
 	private String subscriber_email;
 	
-	public Integer getSubscriber_id() {
+	public Long getSubscriber_id() {
 		return subscriber_id;
 	}
-	public void setSubscriber_id(Integer subscriber_id) {
+	public void setSubscriber_id(Long subscriber_id) {
 		this.subscriber_id = subscriber_id;
 	}
 	public String getFirst_name() {
